+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Create common helper, cover it by RTL/jest tests. UICIRC-803.
 * Cover `LoanPolicy` folder in `Models` by RTL/jest tests. Refs UICIRC-798.
 * Add RTL/Jest testing for `LoanHistoryModel` in `src/settings/Models`. Refs UICIRC-802.
+* Add RTL/Jest testing for `RulesField` component in `src/settings/lib/RuleEditor`. Refs UICIRC-826.
 
 ## [7.0.3](https://github.com/folio-org/ui-circulation/tree/v7.0.3) (2022-04-11)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v7.0.2...v7.0.3)